requirements.txt
# Requirements file for Honeytoken Detection System
# Install with: pip install -r requirements.txt

# Core dependencies
requests>=2.31.0
python-dotenv>=1.0.0

# Optional: For GitHub secret encryption
PyNaCl>=1.5.0

# Development dependencies
pytest>=7.4.0
pytest-cov>=4.1.0

# Code quality
flake8>=6.1.0
black>=23.7.0
mypy>=1.5.0
